Scaffold change

+++ Common/DataTransferObjects/Role/RoleDetail.cs
@@ -2,7 +2,7 @@
 {
     public class RoleDetail
     {
-        public int RoleId { get; set; }
+        public short RoleId { get; set; }
         public string RoleName { get; set; }
         public string RoleDescription { get; set; }
         public bool Active { get; set; }
